add bulk-load with default params
