+++ bench/benches/clustered/benchmark.rs
@@ -4,7 +4,7 @@ use ndarray::{ArrayBase, ArrayView, CowRepr};
 use ordered_float::OrderedFloat;
 use petal_neighbors::BallTree;
 use rand::{rngs::StdRng, seq::SliceRandom, SeedableRng};
-use srtree::{Params, SRTree};
+use srtree::SRTree;
 use std::collections::BinaryHeap;
 
 #[allow(unused)]
@@ -39,7 +39,7 @@ pub fn build_and_query(criterion: &mut Criterion) {
         bencher.iter(|| {
             let pts = benchmark_dataset();
             let pts: Vec<Vec<f64>> = pts.into_iter().map(|p| p.to_vec()).collect();
-            let srtree = SRTree::bulk_load(&pts, Params::default_params());
+            let srtree = SRTree::bulk_load(&pts);
             for point in &query_pts {
                 srtree.query(point, K);
             }
