+++ srtree/tests/test_query.rs
@@ -1,6 +1,6 @@
 use ordered_float::OrderedFloat;
 use rand::prelude::*;
-use srtree::{Params, SRTree};
+use srtree::SRTree;
 
 pub fn euclidean_squared(point1: &[f64], point2: &[f64]) -> f64 {
     if point1.len() != point2.len() {
@@ -15,36 +15,43 @@ pub fn euclidean_squared(point1: &[f64], point2: &[f64]) -> f64 {
 
 #[test]
 fn test_with_random_points() {
-    let number_of_dimensions = 2;
-    let number_of_points = 100;
+    const DIMENSION: usize = 2;
+    let number_of_points = 1000;
     let k = 10;
 
-    let params = Params::new(3, 7, 3, true).unwrap();
-    let mut tree: SRTree<f64> = SRTree::with_params(params);
-
     let mut rng = rand::thread_rng();
-
-    let mut all_points = Vec::new();
-    for i in 0..number_of_points {
+    let mut pts = Vec::new();
+    for _ in 0..number_of_points {
         let mut point_coords = Vec::new();
-        for _ in 0..number_of_dimensions {
-            let x: f64 = 1000000. * rng.gen::<f64>();
+        for _ in 0..DIMENSION {
+            let x: f64 = rng.gen::<f64>();
             point_coords.push(x);
         }
-        tree.insert(&point_coords, i);
-        all_points.push((point_coords, i));
+        pts.push(point_coords);
     }
 
-    let mut points = all_points.clone();
-    for p in all_points.iter() {
-        // SRTree nearest neighbors
-        let result = tree.query(&p.0, k);
+    let mut sequential_tree = SRTree::new();
+    for (index, point) in pts.iter().enumerate() {
+        sequential_tree.insert(point, index);
+    }
+
+    let bulk_tree = SRTree::bulk_load(&pts);
+    let mut points = pts.clone();
+    for p in pts.iter() {
+        // Bulk-loaded SRTree nearest neighbors
+        let result_bulk = bulk_tree.query(p, k);
+        assert_eq!(result_bulk.len(), k);
+
+        // Sequential SRTree nearest neighbors
+        let result_sequential = sequential_tree.query(p, k);
+        assert_eq!(result_sequential.len(), k);
 
         // Brute-force
-        points.sort_by_key(|a| OrderedFloat(euclidean_squared(&a.0, &p.0)));
+        points.sort_by_key(|a| OrderedFloat(euclidean_squared(a, p)));
 
         for i in 0..k {
-            assert_eq!(result[i], points[i].1);
+            assert_eq!(pts[result_bulk[i]], points[i]);
+            assert_eq!(pts[result_sequential[i]], points[i]);
         }
     }
 }
